requirements.txt
python-telegram-bot==21.6
yandex-music==2.2.0
pydantic-settings==2.5.2
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
